evaluation-behaviour-analysis/requirements.txt
Bottleneck==1.3.7
cbor2==5.6.4
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.11.0
fonttools==4.51.0
joblib==1.4.2
kiwisolver==1.4.4
llvmlite==0.43.0
matplotlib==3.9.2
mkl-fft==1.3.8
mkl-random==1.2.4
numba==0.60.0
numexpr==2.8.7
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.2
ply==3.11
psutil==5.9.0
pyarrow==16.1.0
pynndescent==0.5.10
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
setuptools==72.1.0
sip==6.7.12
six==1.16.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.5
tzdata==2023.3
umap-learn==0.5.3
unicodedata2==15.1.0
wheel==0.43.0
